move config and images into src-folder to allow access + change home-component to use new config
