+++ src/img/team/moritzfriedemann.jpg
