--- src/pages/home/config.jsx
@@ -1,94 +1,0 @@
-// Mock Config until database is integrated
-import ProjectImage from "./img/SampleProject.jpg";
-import TeamImage from "./img/SampleTeamMember.png";
-import TeamImage2 from "./img/SampleTeamMember2.png";
-
-export const projectConfig = [
-  {
-    id: "ts-website",
-    name: "Talente Spenden Website",
-    image: ProjectImage,
-    desc: "",
-  },
-  {
-    id: "education",
-    name: "Project: Education",
-    image: ProjectImage,
-    desc: "",
-  },
-  {
-    id: "cfs-website",
-    name: "Coding for Society Website",
-    image: ProjectImage,
-    desc: "The project we made first website",
-  },
-  {
-    id: "hackatum22-0",
-    name: "HackaTUM 2022",
-    image: ProjectImage,
-    desc: "Together we participated in the HackaTUM 2022. We are proud to score the second place in one of the sponsor challenge.",
-  },
-  {
-    id: "hackatum22-1",
-    name: "HackaTUM 2022",
-    image: ProjectImage,
-    desc: "Together we participated in the HackaTUM 2022. We are proud to score the second place in one of the sponsor challenge.",
-  },
-  {
-    id: "hackatum22-2",
-    name: "HackaTUM 2022",
-    image: ProjectImage,
-    desc: "Together we participated in the HackaTUM 2022. We are proud to score the second place in one of the sponsor challenge.",
-  },
-  {
-    id: "hackatum22-3",
-    name: "HackaTUM 2022",
-    image: ProjectImage,
-    desc: "Together we participated in the HackaTUM 2022. We are proud to score the second place in one of the sponsor challenge.",
-  },
-];
-
-export const teamConfig = [
-  {
-    id: "jonasitt",
-    name: "Jonas Itt",
-    role: "Founder",
-    image: TeamImage,
-    desc: "I want to use my skills for the good. This is why...",
-  },
-  {
-    id: "nickalexanderhuebner",
-    name: "Nick Alexander Hübner",
-    role: "Founder",
-    image: TeamImage2,
-    desc: "My motivation is to gain experience and do you good at the same ...",
-  },
-  {
-    id: "jakobguenther",
-    name: "Jakob Günther",
-    role: "Team Member",
-    image: TeamImage,
-    desc: "My motivation is to gain experience and do you good at the same ...",
-  },
-  {
-    id: "nhatnguyen",
-    name: "Nhat Nguyen",
-    role: "Founder",
-    image: TeamImage,
-    desc: "My motivation is to gain experience and do you good at the same ...",
-  },
-  {
-    id: "haoruiharryzheng",
-    name: "Haorui Harry Zheng",
-    role: "Founder",
-    image: TeamImage,
-    desc: "My motivation is to gain experience and do you good at the same ...",
-  },
-  {
-    id: "daniellindner",
-    name: "Daniel Linder",
-    role: "Team Member",
-    image: TeamImage,
-    desc: "My motivation is to gain experience and do you good at the same ...",
-  },
-];
